fix(app): various fixes and improvements

+++ app/api/board/delete/route.ts
@@ -5,6 +5,13 @@ export const POST = async (req: Request) => {
   try {
     const { boardId } = await new Response(req.body).json();
 
+    if (!boardId) {
+      return NextResponse.json(
+        { error: `The required body param 'boardId' was not provided` },
+        { status: 422 },
+      );
+    }
+
     await sql`DELETE FROM Boards WHERE id = ${boardId} RETURNING *;`;
 
     return NextResponse.json({ boardId });
